+++ Source/Coco.Server/Coco.Entities/Dtos/Auth/AuthorizationPolicyRolesDto.cs
@@ -4,7 +4,7 @@ namespace Coco.Entities.Dtos.Auth
 {
     public class AuthorizationPolicyRolesDto
     {
-        public short Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
